+++ src/test/scala/sangria/integration/SprayJsonSupportSpec.scala
@@ -2,7 +2,7 @@ package sangria.integration
 
 import org.scalatest.{Matchers, WordSpec}
 import sangria.execution.Executor
-import sangria.integration.SprayJsonSupport.SprayJsonResultMarshaller
+import sangria.integration.sprayJson.SprayJsonResultMarshaller
 import sangria.parser.QueryParser
 import sangria.starWars.TestData.{CharacterRepo, FriendsResolver}
 import sangria.starWars.TestSchema._
@@ -16,7 +16,7 @@ import scala.util.Success
 class SprayJsonSupportSpec extends WordSpec with Matchers with AwaitSupport {
   "Json4sSupport" should {
     "Marshal and Unmarshal" in {
-      import sangria.integration.SprayJsonSupport._
+      import sangria.integration.sprayJson._
 
       val Success(query) = QueryParser.parse("""
         query FetchSomeIDQuery($someId: String!) {